Updating "Aliquota" type and fixing error logging logic

+++ Lacuna.FocusNFSeIntegration/Models/ServiceInfo.cs
@@ -8,7 +8,7 @@ namespace Lacuna.FocusNFSeIntegration.Models {
 	public class ServiceInfo {
 
 		[JsonProperty("aliquota")]
-		public int Aliquota { get; set; }
+		public double Aliquota { get; set; }
 
 		[JsonProperty("discriminacao")]
 		public string Description { get; set; }
